ci: add Makefile with ci/run targets; fix 5 SpotBugs bugs

+++ src/main/java/ui/controller/admin/BookingManagmentController.java
@@ -63,7 +63,7 @@ public class BookingManagmentController {
             showAlert("No Selection", "Please select a booking to cancel.");
             return;
         }
-        ctx.bookingService.cancelBooking((long) selected.getBookingId());
+        ctx.bookingService.cancelBooking(selected.getBookingId());
         showAlert("Success", "Booking Cancelled");
         loadTableData();
     }
